requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence-transformers
numpy
